refactor(bot): finish skill_orchestration extraction — closes #119
Completes the #119 reduced-scope follow-up from PR #122. Migrates the
remaining 16 internal helpers (~620 LOC) from ``bot/core.py`` into
``bot/skill_orchestration.py`` and re-routes
``omicsclaw/runtime/preflight/sc_batch.py`` to import them from their
new canonical home (PR #122 review Important #2).

Functions moved (16 total):

* Memory auto-capture: ``_auto_capture_dataset``,
  ``_auto_capture_analysis``.
* Env-error parsing: ``_extract_env_snippet``, ``_extract_fix_hint``,
  ``_classify_env_error``, plus the ``_ENV_ERROR_PATTERNS`` table.
* Output state helpers: ``_resolve_last_output_dir``,
  ``_read_result_json``, ``_update_preprocessing_state``,
  ``_format_next_steps``.
* Skill execution: ``_normalize_extra_args``, ``_run_omics_skill_step``,
  ``_run_skill_via_shared_runner``.
* Skill lookup + param hints: ``_lookup_skill_info``,
  ``_resolve_param_hint_info``, ``_infer_skill_for_method``,
  ``_build_method_preview``, ``_build_param_hint``.

Cross-module concerns (late imports inside the moved functions):

* ``_auto_capture_dataset`` / ``_auto_capture_analysis`` /
  ``_resolve_last_output_dir`` / ``_update_preprocessing_state``
  late-import ``memory_store`` and ``OMICSCLAW_DIR`` from ``bot.core``.
* ``_run_omics_skill_step`` late-imports ``OUTPUT_DIR`` from ``bot.core``.
* ``_infer_skill_for_method`` late-imports ``_skill_registry`` from
  ``bot.core``.
* ``omicsclaw/runtime/preflight/sc_batch.py`` now imports
  ``_lookup_skill_info`` and ``_run_omics_skill_step`` from
  ``bot.skill_orchestration`` directly. ``execute_omicsclaw`` still
  comes from ``bot.core`` (will move to ``bot.tool_executors`` in
  #120; this PR's second commit).

Backward-compat:

* ``bot.core`` re-exports every previously-public symbol via two new
  re-export blocks. External callers continue to find
  ``bot.core._run_omics_skill_step``, ``bot.core._classify_env_error``,
  etc. unchanged.

Tests added:

* ``tests/bot/test_skill_orchestration.py`` — single comprehensive
  identity-comparison test enforcing the backward-compat contract for
  all 22 symbols on ``bot.skill_orchestration`` (5 from PR #122 + 17
  this slice). Forces parallel-copy detection: ``bot.core.X is
  bot.skill_orchestration.X`` for every name.

Verification:

* Bot regression net: 226 passed (+1 new identity test), 1 skipped,
  4 xfailed — no regressions.
* ``tests/eval/`` + ``tests/test_knowhow.py``: 121 passed, 17 deselected.
  Eval baseline 14/15 against ``deepseek-v4-pro`` unaffected.
* ``bot/core.py``: 3811 → 3191 LOC (−620). Cumulative drop from main
  baseline: 5160 → 3191 (−1969, −38.2%).
* ``bot/skill_orchestration.py``: 116 → 771 LOC.

Closes #119.
